save my fixes

+++ classes_and_utils/TemplatesFilesHelper.py
@@ -1,8 +1,68 @@
 import sys, os, json
 from glob import glob
 from pathlib import Path
+from jsonschema import validate
+import jsonschema
+
 
 REPORTS_TEMPLATES_FOLDER_NAME = "reports_templates"
+
+schema = {
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "type": "object",
+  "properties": {
+    "name": {
+      "type": "string"
+    },
+    "segmentations": {
+      "type": "array",
+      "items": [
+        {
+          "type": "object",
+          "properties": {
+            "name": {
+              "type": "string"
+            },
+            "rows": {
+              "type": "string"
+            },
+            "columns": {
+              "type": "string"
+            }
+          },
+          "required": [
+            "name",
+            "rows",
+            "columns"
+          ]
+        },
+        {
+          "type": "object",
+          "properties": {
+            "name": {
+              "type": "string"
+            },
+            "rows": {
+              "type": "string"
+            },
+            "columns": {
+              "type": "string"
+            }
+          },
+          "required": [
+            "name",
+            "rows",
+            "columns"
+          ]
+        }
+      ]
+    }
+  },
+  "required": [
+    "name",
+    "segmentations"
+  ]
+}
 
 class TemplateItem:
     name = ''
@@ -14,22 +74,8 @@ class TemplateItem:
 
 class TemplatesFilesHelper:
     
-    def __init__(self):
-    
-        pass
-
-    def get_all_templates_names(self):
-        templates = []
-        path = str(os.path.join(str(Path(os.path.dirname(os.path.realpath(__file__))).parent), REPORTS_TEMPLATES_FOLDER_NAME,'*'))
-        files = glob(path)
-        for fullname in files:
-            filename = fullname.split(os.sep)[-1]
-            file_parts = filename.split('.')
-            if len(file_parts) == 2 and file_parts[1] == 'json':
-                templates.append(file_parts[0])
-        return templates
-    
     def save_template(self,filename,content,main_path,ref_path,returnAllTemplates = True):
+        
         _, file_extension = os.path.splitext(filename)
         if file_extension == '' or file_extension == None:
             filename += ".json"
@@ -39,8 +85,9 @@ class TemplatesFilesHelper:
         dir_main,_ = os.path.split(main_path)
         path = os.path.join(dir_main,filename)
         
-        with open(path,'w') as f:
-            f.write(content)
+        if os.path.exists(dir_main):
+            with open(path,'w') as f:
+                f.write(content)
 
         dir_ref,_ = os.path.split(ref_path)
 
@@ -79,7 +126,11 @@ class TemplatesFilesHelper:
         files = glob(folder + '\*.json', recursive=recursive)
         for fullname in files:
             with open(fullname,'r') as f:
-                content = json.load(f)  
-                    
-            templates.append({'name':os.path.splitext(os.path.basename(fullname))[0],'content':content})
+                try:
+                    content = json.load(f)
+                    validate(instance=content, schema=schema)
+                    templates.append({'name':os.path.splitext(os.path.basename(fullname))[0],'content':content})
+                except jsonschema.exceptions.ValidationError as err:
+                    print(err)
+                      
         return templates
